Begin split-up and rewrite of codebase
